Change osu!catch juice stream tick generation to match osu!(stable)
Closes #20508.

+++ osu.Game.Rulesets.Catch/Objects/JuiceStream.cs
@@ -70,7 +70,8 @@ namespace osu.Game.Rulesets.Catch.Objects
                 // generate tiny droplets since the last point
                 if (lastEvent != null)
                 {
-                    double sinceLastTick = e.Time - lastEvent.Value.Time;
+                    // int truncation is used here intentionally to match stable's generation logic generation.
+                    int sinceLastTick = (int)(e.Time - lastEvent.Value.Time);
 
                     if (sinceLastTick > 80)
                     {
